refactor: move SumWeightTool to tools folder
