+++ src/physana/tools/__init__.py
@@ -1,5 +1,12 @@
 from .xsec import PMGXsec
 from .file_metadata import FileMetaData
 from .data_file_check import check_data_completeness
+from .sum_weights import SumWeightTool, extract_cutbook_sum_weights
 
-__all__ = ['PMGXsec', 'FileMetaData', 'check_data_completeness']
+__all__ = [
+    'PMGXsec',
+    'FileMetaData',
+    'check_data_completeness',
+    'SumWeightTool',
+    'extract_cutbook_sum_weights',
+]
