+++ src/physana/tools/sum_weights.py
